backend/requirements.txt
Django
djangorestframework
djoser

python-dotenv

Pillow

django-storages
boto3

psycopg2-binary

drf-spectacular

gunicorn

redis
celery[redis]
django-celery-beat
django-redis

django-debug-toolbar
